Update function declarations in src/codec.h and update conversions

+++ src/codec.h
@@ -19,8 +19,8 @@ long double codec_linear_l_from_float64(double);
 long double codec_linear_l_from_extended_float(long double);
 
 takum8 codec_takum8_from_s_and_l(bool, float);
-takum16 codec_takum16_from_s_and_l(bool, float);
-takum32 codec_takum32_from_s_and_l(bool, double);
+takum16 codec_takum16_from_s_and_l(bool, double);
+takum32 codec_takum32_from_s_and_l(bool, long double);
 takum64 codec_takum64_from_s_and_l(bool, long double);
 
 #endif /* CODEC_H */
